// account-system/src/app/models/CheckBook.ts
// Generated by https://quicktype.io

export interface CheckBook {
    dbid:              number;
    accountId:         number;
    start_correlative: number;
    fin_correlative:   number;
    quantity_checks:   number;
    status:            number;
    user_registry:     string;
    user_modified:     string;
    registry_date:      Date;
}